Refactor code to meet .NET6 standards, turn some frequently used values into properties, move unused code into obsolete file
