+++ EpiResponsivePicture/ResizedImage/Property/FocalPoint.cs
@@ -14,6 +14,8 @@ namespace Forte.EpiResponsivePicture.ResizedImage.Property
         public double X { get; }
         public double Y { get; }
 
+        public static FocalPoint Center => new(0.5, 0.5);
+
         public override string ToString()
         {
             return $"{X.ToString(CultureInfo.InvariantCulture)}|{Y.ToString(CultureInfo.InvariantCulture)}";
